- made output more compliant with old version - added PerSCorer.h and BleuScorer.h - stored feature names - fixed bug about output of best Point
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk@1796 1f5c12ca-751b-0410-a591-d2e778427230

+++ mert/PerScorer.cpp
@@ -1,4 +1,4 @@
-#include "Scorer.h"
+#include "PerScorer.h"
 
 
 void PerScorer::setReferenceFiles(const vector<string>& referenceFiles) {
@@ -32,7 +32,7 @@ void PerScorer::setReferenceFiles(const vector<string>& referenceFiles) {
 
 }
 
-void PerScorer::prepareStats(unsigned int sid, const string& text, ScoreStats& entry) {
+void PerScorer::prepareStats(size_t sid, const string& text, ScoreStats& entry) {
     if (sid >= _reflengths.size()) {
         stringstream msg;
         msg << "Sentence id (" << sid << ") not found in reference set";
